Move test files into tests folder.
